Исправления замечаний #1 * Добавил новый обработчик исключений для исключений любого типа, со статусом 500. * Добавил лидирующий слэш для аннотаций PutMapping и DeleteMapping. * Класс ErrorResponse перенес в ErrorHandler. * Переименовал интерфейсы и классы сервисов. * Изменил сортировку в stream у метода findPopularMovies в FilmService, и так же добавил в метод проверку count на положительное число. * Переименовал константу DATE_MOVIES_BIRTHDAY класса MovieReleaseDateValidator на MOVIES_BIRTHDAY. * Переименовал параметры методов getCommonFriends, getFriends, deleteFriend, addFriend класса UserServiceImpl. * Переименовал параметры методов addLike, deleteLike класса FilmServiceImpl.
