--- src/main/java/ru/yandex/practicum/filmorate/model/ErrorResponse.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-import lombok.Getter;
-
-@Getter
-public class ErrorResponse {
-    private final String error;
-
-    public ErrorResponse(String error) {
-        this.error = error;
-    }
-}
